server/requirements.txt
flask
flask-cors
awesome-slugify

